:rocket Made the code modular
