+++ pages/Purchase.jsx
@@ -17,6 +17,7 @@ import { GrandCalculations } from "../components/GrandCalculations";
 import { TableHeader } from "../components/TableHeader";
 import * as MediaLibrary from "expo-media-library";
 import { captureRef } from "react-native-view-shot";
+import {calculateCft, categorize, categoryTotalCFT, getTotalPrice, onLayout, onSaveImageAsync} from "./utils/Purchase";
 
 const height = Dimensions.get("window").height;
 const width = Dimensions.get("window").width;
@@ -96,20 +97,6 @@ export const Purchase = ({ navigation }) => {
     return () => backHandler.remove();
   }, []);
 
-  const calculateCft = () => {
-    const CFT = (length * girth * girth) / 2304;
-    setList((prev) => {
-      return [
-        ...prev,
-        {
-          length: length,
-          girth: girth,
-          CFT: CFT,
-        },
-      ];
-    });
-  };
-
   const storeListInAsyncStorage = async (key, value) => {
     try {
       await AsyncStorage.setItem(key, JSON.stringify(value));
@@ -139,184 +126,12 @@ export const Purchase = ({ navigation }) => {
     }
   };
 
-  const categorize = () => {
-    for (let i = 0; i < list.length; i++) {
-      if (list[i].girth < 12 && list[i].girth > 0) {
-        setOneToEleven((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth < 18) {
-        setTwelveToSeventeen((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth < 24) {
-        setEighteenToTwentyThree((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth < 30) {
-        setTwentyFourToTwentyNine((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth < 36) {
-        setThirtyToThirtyFive((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth < 48) {
-        setThirtySixToFortySeven((prev) => {
-          return [...prev, list[i]];
-        });
-      } else if (list[i].girth >= 48) {
-        setFortyEightAbove((prev) => {
-          return [...prev, list[i]];
-        });
-      }
-    }
-  };
-
-  const categoryTotalCFT = () => {
-    let temp = 0;
-    let temp2 = 0;
-    let temp3 = 0;
-    let temp4 = 0;
-    let temp5 = 0;
-    let temp6 = 0;
-    let temp7 = 0;
-    for (let i = 0; i < oneToEleven.length; i++) {
-      temp += +oneToEleven[i].CFT;
-    }
-    for (let i = 0; i < TwelveToSeventeen.length; i++) {
-      temp2 += +TwelveToSeventeen[i].CFT;
-    }
-    for (let i = 0; i < EighteenToTwentyThree.length; i++) {
-      temp3 += +EighteenToTwentyThree[i].CFT;
-    }
-    for (let i = 0; i < TwentyFourToTwentyNine.length; i++) {
-      temp4 += +TwentyFourToTwentyNine[i].CFT;
-    }
-    for (let i = 0; i < ThirtyToThirtyFive.length; i++) {
-      temp5 += +ThirtyToThirtyFive[i].CFT;
-    }
-    for (let i = 0; i < ThirtySixToFortySeven.length; i++) {
-      temp6 += +ThirtySixToFortySeven[i].CFT;
-    }
-    for (let i = 0; i < FortyEightAbove.length; i++) {
-      temp7 += +FortyEightAbove[i].CFT;
-    }
-    setOne(temp);
-    console.log(temp);
-    console.log(one);
-    setTwo(temp2);
-    setThree(temp3);
-    setFour(temp4);
-    setFive(temp5);
-    setSix(temp6);
-    setSeven(temp7);
-  };
-
   useEffect(() => {
-    getTotalPrice();
+    getTotalPrice(one, two, three, four, five, six, seven, setGrandPrice, setGrandCFT, oneToElevenPrice, TwelveToSeventeenPrice, EighteenToTwentyThreePrice, TwentyFourToTwentyNinePrice, ThirtyToThirtyFivePrice, ThirtySixToFortySevenPrice, FortyEightAbovePrice);
   }, [one]);
-
-  const getTotalPrice = () => {
-    const onePrice = parseInt(one) * parseInt(oneToElevenPrice);
-    console.log(`total cft: ${one}`);
-    const twoPrice = parseInt(two) * parseInt(TwelveToSeventeenPrice);
-    const threePrice = parseInt(three) * parseInt(EighteenToTwentyThreePrice);
-    const fourPrice = parseInt(four) * parseInt(TwentyFourToTwentyNinePrice);
-    const fivePrice = parseInt(five) * parseInt(ThirtyToThirtyFivePrice);
-    const sixPrice = parseInt(six) * parseInt(ThirtySixToFortySevenPrice);
-    const sevenPrice = parseInt(seven) * parseInt(FortyEightAbovePrice);
-    const isValidNumber = (value) => typeof value === "number" && !isNaN(value);
-
-    // Calculate the total price, considering undefined variables as 0
-    const totalPrice =
-      (onePrice ? onePrice : 0) +
-      (twoPrice ? twoPrice : 0) +
-      (threePrice ? threePrice : 0) +
-      (fourPrice ? fourPrice : 0) +
-      (fivePrice ? fivePrice : 0) +
-      (sixPrice ? sixPrice : 0) +
-      (sevenPrice ? sevenPrice : 0);
-    setGrandPrice(totalPrice);
-
-    // Calculate the total CFT, considering undefined variables as 0
-    const totalCFT =
-      (one ? one : 0) +
-      (two ? two : 0) +
-      (three ? three : 0) +
-      (four ? four : 0) +
-      (five ? five : 0) +
-      (six ? six : 0) +
-      (seven ? seven : 0);
-    setGrandCFT(totalCFT);
-  };
 
   // get height of table view
   const [tableHeight, setTableHeight] = useState(0);
-  const onLayout = (event) => {
-    const { x, y, height, width } = event.nativeEvent.layout;
-    setTableHeight(height);
-  };
-
-  // save view as local media in the device
-  const onSaveImageAsync = async () => {
-    console.log(`tableHeight: ${tableHeight}`);
-    try {
-      const localUri = await captureRef(imageRef, {
-        height: tableHeight,
-        quality: 1,
-      });
-
-      await MediaLibrary.saveToLibraryAsync(localUri);
-      console.log(localUri);
-      await saveFile(localUri);
-      if (localUri) {
-        alert("Saved!");
-      }
-    } catch (e) {
-      console.log(e);
-    }
-  };
-
-  // adding to auto print album
-  async function saveFile(filePath) {
-    const albumName = "auto print";
-    const permission = await MediaLibrary.requestPermissionsAsync();
-
-    let asset = null;
-    if (permission.granted) {
-      try {
-        asset = await MediaLibrary.createAssetAsync(filePath);
-      } catch (e) {
-        console.error("MediaLibrary.createAssetAsync failed", e);
-      }
-
-      if (asset) {
-        try {
-          let album = await MediaLibrary.getAlbumAsync(albumName);
-          if (album) {
-            await MediaLibrary.addAssetsToAlbumAsync([asset], album, false);
-          } else {
-            album = await MediaLibrary.createAlbumAsync(
-              albumName,
-              asset,
-              false
-            );
-          }
-          const assetResult = await MediaLibrary.getAssetsAsync({
-            first: 1,
-            album,
-            sortBy: MediaLibrary.SortBy.creationTime,
-          });
-          asset = await assetResult.assets[0];
-        } catch (e) {
-          console.error(" failed", e);
-        }
-      } else {
-        console.error("unable to use MediaLibrary, can not create assets");
-      }
-    }
-  }
 
   return (
     <ScrollView style={styles.parentView}>
@@ -340,7 +155,7 @@ export const Purchase = ({ navigation }) => {
               placeholder="Girth"
               keyboardType="decimal-pad"
               onSubmitEditing={async () => {
-                await calculateCft();
+                await calculateCft(length, girth, setList);
                 await lengthRef.current.focus();
                 setLength();
                 setGirth();
@@ -362,7 +177,7 @@ export const Purchase = ({ navigation }) => {
                   text: "Ok",
                   style: "OK",
                   onPress: () => {
-                    categorize();
+                    categorize(list, setOneToEleven, setTwelveToSeventeen, setEighteenToTwentyThree, setTwentyFourToTwentyNine, setThirtyToThirtyFive, setThirtySixToFortySeven, setFortyEightAbove);
                     setIsPriceScreen(true);
                   },
                 },
@@ -541,9 +356,9 @@ export const Purchase = ({ navigation }) => {
           <TouchableOpacity
           style={styles.addPrice}
             onPress={() => {
-              categoryTotalCFT();
+              categoryTotalCFT(oneToEleven, TwelveToSeventeen, EighteenToTwentyThree, TwentyFourToTwentyNine, ThirtyToThirtyFive, ThirtySixToFortySeven, FortyEightAbove, setOne, setTwo, setThree, setFour, setFive, setSix, setSeven);
               // getCategoryPrice();
-              getTotalPrice();
+              getTotalPrice(one, two, three, four, five, six, seven, setGrandPrice, setGrandCFT, oneToElevenPrice, TwelveToSeventeenPrice, EighteenToTwentyThreePrice, TwentyFourToTwentyNinePrice, ThirtyToThirtyFivePrice, ThirtySixToFortySevenPrice, FortyEightAbovePrice);
               setIsInvoiceScreen(true);
               setIsPriceScreen(false);
             }}
@@ -554,7 +369,7 @@ export const Purchase = ({ navigation }) => {
         </View>
       )}
       {isInvoiceScreen && (
-        <View ref={imageRef} onLayout={onLayout} style={{ overflow: "hidden" }}>
+        <View ref={imageRef} onLayout={(event) => onLayout(event, setTableHeight)} style={{ overflow: "hidden" }}>
           <Text> Date: Time:</Text>
           <TableHeader />
           {oneToEleven.length > 0 && (
@@ -613,7 +428,7 @@ export const Purchase = ({ navigation }) => {
           tableIndex={tableIndex} setTableIndex={setTableIndex} />
           <TouchableOpacity
             style={styles.printButton}
-            onPress={onSaveImageAsync}
+            onPress={() => onSaveImageAsync(tableHeight, imageRef)}
           >
             <Text style={styles.addPriceText}>
               Print
